# docs/swagger.yaml
basePath: /api
definitions:
  course.RequestDto:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  course.ResponseDto:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  hubspot.UserAssociation:
    properties:
      results:
        items:
          properties:
            id:
              type: string
            type:
              type: string
          type: object
        type: array
    type: object
  hubspot.UserAssociations:
    properties:
      contacts:
        $ref: '#/definitions/hubspot.UserAssociation'
    type: object
  hubspot.UserProps:
    properties:
      email:
        type: string
      firstname:
        type: string
      lastname:
        type: string
    type: object
  hubspot.UserResponse:
    properties:
      associations:
        $ref: '#/definitions/hubspot.UserAssociations'
      createdAt:
        type: string
      id:
        type: string
      properties:
        $ref: '#/definitions/hubspot.UserProps'
      updatedAt:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
paths:
...
  /courses:
    get:
      consumes:
      - application/json
      description: Get a list of courses
      parameters:
      - description: Filter by course name
        in: query
        name: name
        type: string
      - description: Filter by course description
        in: query
        name: description
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetMemberships of courses retrieved successfully
          schema:
            items:
              $ref: '#/definitions/course.ResponseDto'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get a list of courses
      tags:
      - courses
    post:
      consumes:
      - application/json
      description: Create a new course
      parameters:
      - description: Course details
        in: body
        name: course
        required: true
        schema:
          $ref: '#/definitions/course.RequestDto'
      produces:
      - application/json
      responses:
        "201":
          description: Course created successfully
          schema:
            $ref: '#/definitions/course.ResponseDto'
        "400":
          description: 'Bad Request: Invalid input'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Create a new course
      tags:
      - courses
  /courses/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a course by HubSpotId
      parameters:
      - description: Course HubSpotId
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: 'Bad Request: Invalid HubSpotId'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'Not Found: Course not found'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Delete a course
      tags:
      - courses
    get:
      consumes:
      - application/json
      description: Get a course by HubSpotId
      parameters:
      - description: Course HubSpotId
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Course retrieved successfully
          schema:
            $ref: '#/definitions/course.ResponseDto'
        "400":
          description: 'Bad Request: Invalid HubSpotId'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'Not Found: Course not found'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get a course by HubSpotId
      tags:
      - courses
    put:
      consumes:
      - application/json
      description: Update a course
      parameters:
      - description: Course HubSpotId
        in: path
        name: id
        required: true
        type: string
      - description: Course details
        in: body
        name: course
        required: true
        schema:
          $ref: '#/definitions/course.RequestDto'
      produces:
      - application/json
      responses:
        "204":
          description: 'No Content: Course updated successfully'
        "400":
          description: 'Bad Request: Invalid input'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'Not Found: Course not found'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Update a course
      tags:
      - courses
  /customers:
    get:
      consumes:
      - application/json
      description: Retrieves customers based on an event HubSpotId or returns all
        customers
      parameters:
      - description: Event HubSpotId (if specified, fetches customers for the event)
        in: query
        name: event_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Customers retrieved successfully
          schema:
            items:
              $ref: '#/definitions/hubspot.UserResponse'
            type: array
        "400":
          description: 'Bad Request: Invalid parameters'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get customers
      tags:
      - customers
  /customers/{email}:
    get:
      consumes:
      - application/json
      description: Retrieves a customer using their email address
      parameters:
      - description: Customer Email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Customer retrieved successfully
          schema:
            $ref: '#/definitions/hubspot.UserResponse'
        "400":
          description: 'Bad Request: Invalid Email'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'Not Found: Customer not found'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get a customer by email
      tags:
      - customers
  /customers/{email}/children:
    get:
      consumes:
      - application/json
      description: Retrieves a customer's children using the parent's email address
      parameters:
      - description: Parent Email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Customer's children retrieved successfully
          schema:
            items:
              $ref: '#/definitions/hubspot.UserResponse'
            type: array
        "400":
          description: 'Bad Request: Invalid Email'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'Not Found: Parent or children not found'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get a customer's children by parent email
      tags:
      - customers
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"